fix(achievements): scope the stats header to the active tab
